+++ dutchiepay/src/app/_components/Logo.jsx
@@ -1,8 +1,8 @@
-'use client';
-
 import '@/styles/globals.css';
 
-import logo from '../../../../public/image/logo.jpg';
+import Image from 'next/image';
+import Link from 'next/link';
+import logo from '../../../public/image/logo.jpg';
 
 export default function Logo() {
   return (
