+++ dutchiepay/src/app/_components/_user/SocialLogin.jsx
@@ -1,3 +1,5 @@
+'use client';
+
 import '@/styles/globals.css';
 import '@/styles/user.css';
